+++ R/Swim_Parse.R
@@ -110,11 +110,11 @@ Swim_Parse <-
       stop("typo and replacement must have the same number of elements (be the same length)")
     }
 
-    if(any(!is.logical(format_results) & is.na(format_results)) == TRUE) {
+    if(any(!is.logical(format_results), is.na(format_results)) == TRUE) {
       stop("format_results must be logical, either TRUE or FALSE")
     }
 
-    if(any(!is.logical(splits) & is.na(splits)) == TRUE) {
+    if(any(!is.logical(splits), is.na(splits)) == TRUE) {
       stop("splits must be logical, either TRUE or FALSE")
     }
 
@@ -122,7 +122,7 @@ Swim_Parse <-
       stop("split_length must be numeric, usually 50 or 25")
     }
 
-    if(any(!is.logical(relay_swimmers) & is.na(relay_swimmers)) == TRUE) {
+    if(any(!is.logical(relay_swimmers), is.na(relay_swimmers)) == TRUE) {
       stop("relay_swimmers must be logical, either TRUE or FALSE")
     }